create useSitePreviewShareCode hook
