+++ client/landing/stepper/declarative-flow/internals/steps-repository/launchpad/launchpad-site-preview.tsx
@@ -1,16 +1,9 @@
-import { isWpComBusinessPlan, isWpComEcommercePlan } from '@automattic/calypso-products';
 import { DEVICE_TYPES } from '@automattic/components';
 import { NEWSLETTER_FLOW, VIDEOPRESS_FLOW } from '@automattic/onboarding';
 import { addQueryArgs } from '@wordpress/url';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect } from 'react';
-import { useCreateSitePreviewLink } from 'calypso/components/site-preview-link/use-create-site-preview-link';
-import {
-	PreviewLink,
-	useSitePreviewLinks,
-} from 'calypso/components/site-preview-link/use-site-preview-links';
 import WebPreview from 'calypso/components/web-preview/component';
-import { useSite } from 'calypso/landing/stepper/hooks/use-site';
+import { useSitePreviewShareCode } from 'calypso/landing/stepper/hooks/use-site-preview-share-code';
 import { usePremiumGlobalStyles } from 'calypso/state/sites/hooks/use-premium-global-styles';
 import PreviewToolbar from '../design-setup/preview-toolbar';
 import type { Device } from '@automattic/components';
@@ -22,7 +15,6 @@ const LaunchpadSitePreview = ( {
 	siteSlug: string | null;
 	flow: string | null;
 } ) => {
-	const site = useSite();
 	const translate = useTranslate();
 	const { globalStylesInUse, shouldLimitGlobalStyles } = usePremiumGlobalStyles();
 
@@ -30,49 +22,14 @@ const LaunchpadSitePreview = ( {
 	const devicesToShow: Device[] = [ DEVICE_TYPES.COMPUTER, DEVICE_TYPES.PHONE ];
 	let defaultDevice = flow === NEWSLETTER_FLOW ? DEVICE_TYPES.COMPUTER : DEVICE_TYPES.PHONE;
 	const isVideoPressFlow = VIDEOPRESS_FLOW === flow;
-	const isBusinessPlan = site?.plan?.product_slug
-		? isWpComBusinessPlan( site?.plan?.product_slug )
-		: false;
-	const isEcommercePlan = site?.plan?.product_slug
-		? isWpComEcommercePlan( site?.plan?.product_slug )
-		: false;
 
 	if ( isVideoPressFlow ) {
 		const windowWidth = window.innerWidth;
 		defaultDevice = windowWidth >= 1000 ? DEVICE_TYPES.COMPUTER : DEVICE_TYPES.PHONE;
 	}
 
-	const usePreviewSiteLinksQueryEnabled =
-		site?.is_coming_soon && ( isBusinessPlan || isEcommercePlan ) && site?.is_wpcom_atomic;
-
-	const { data: previewLinks, isLoading: isPreviewLinksLoading } = useSitePreviewLinks( {
-		siteId: Number( site?.ID ),
-		isEnabled: usePreviewSiteLinksQueryEnabled ?? false,
-	} );
-
-	const { createLink, isLoading: isCreatingSitePreviewLinks } = useCreateSitePreviewLink( {
-		siteId: Number( site?.ID ),
-	} );
-
-	// Generate preview link for site on business or ecommerce plan
-	// Preview links are only available on these two plans
-	useEffect( () => {
-		if ( previewLinks && Array.isArray( previewLinks ) && previewLinks.length === 0 ) {
-			if ( isBusinessPlan || isEcommercePlan ) {
-				createLink();
-			}
-		}
-	}, [ previewLinks, createLink, isBusinessPlan, isEcommercePlan ] );
-
-	const shareCode = getPreviewCode( previewLinks );
-
-	function getPreviewCode( links: PreviewLink[] | undefined ) {
-		if ( typeof links === 'undefined' ) {
-			return false;
-		}
-		const link = links[ 0 ];
-		return link?.code ?? false;
-	}
+	const { shareCode, isPreviewLinksLoading, isCreatingSitePreviewLinks } =
+		useSitePreviewShareCode();
 
 	function formatPreviewUrl() {
 		if ( ! previewUrl || isPreviewLinksLoading || isCreatingSitePreviewLinks ) {
